+++ packages/server/package.json
@@ -4,9 +4,9 @@
   "description": "",
   "scripts": {
     "preview": "node ./dist/index.js",
+    "build": "rimraf -rf dist && tsc --p ./tsconfig.prod.json",
     "dev": "cross-env NODE_ENV=development nodemon index.ts",
-    "serve": "cross-env NODE_ENV=production node server",
-    "serve dev": "node server",
+    "serve": "cross-env NODE_ENV=production node dist/index.js",
     "lint": "eslint .",
     "format": "prettier --write .",
     "test": "jest ."
@@ -29,6 +29,7 @@
     "@types/jest": "29.4.0",
     "@types/node": "18.14.6",
     "@types/pg": "8.6.6",
+    "@types/serve-static": "^1.15.1",
     "@typescript-eslint/eslint-plugin": "5.54.0",
     "@typescript-eslint/parser": "5.54.0",
     "@vitejs/plugin-react": "3.1.0",
@@ -42,6 +43,7 @@
     "jest": "29.4.3",
     "nodemon": "2.0.21",
     "prettier": "2.8.4",
+    "rimraf": "^4.4.0",
     "serve-static": "^1.15.0",
     "ts-jest": "29.0.5",
     "ts-node": "10.9.1",
