+++ packages/server/tsconfig.json
@@ -2,8 +2,9 @@
   "ts-node": { "esm": true },
   "compilerOptions": {
     "composite": true,
-    "module": "NodeNext",
-    "moduleResolution": "nodenext",
+    "target": "esnext",
+    "module": "esnext",
+    "moduleResolution": "Node",
     "esModuleInterop": true,
     "sourceMap": true,
     "declaration": true,
@@ -24,6 +25,6 @@
     "paths": {
       "*": ["types/*"]
     },
-    "outDir": "dist"
+    "outDir": "dist",
   }
 }
